# cat-service/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Welcome to cat-service.",
      "version": "1.0.0",
      "title": "cat-service"
    },
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/devices/{id}/speakers": {
            "get": {
                "tags": [
                    "speakers"
                ],
                "description": "Fetch speakers by device id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device id",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found speakers",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/speaker"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/speakers": {
            "get": {
                "tags": [
                    "speakers"
                ],
                "description": "Fetch speakers",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "afterDate",
                        "in": "query",
                        "description": "Fetch speakers created after the provided date",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "beforeDate",
                        "in": "query",
                        "description": "Fetch speakers created before the provided date",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found speakers",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/speaker"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "speakers"
                ],
                "description": "Create speaker",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "formData",
                        "description": "Device used to identify speaker",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "data",
                        "in": "formData",
                        "description": "Speaker data",
                        "required": true,
                        "type": "string",
                        "format": "json"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Speaker created",
                        "schema": {
                            "$ref": "#/definitions/speaker"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/speakers/{id}": {
            "get": {
                "tags": [
                    "speakers"
                ],
                "description": "Fetch speaker",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device id",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found speaker",
                        "schema": {
                            "$ref": "#/definitions/speaker"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "speakers"
                ],
                "description": "Delete speaker",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Speaker id",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Speaker deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "speakers"
                ],
                "description": "Update speaker by id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Speaker id",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "data",
                        "in": "formData",
                        "description": "Speaker data",
                        "required": true,
                        "type": "string",
                        "format": "json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated speaker",
                        "schema": {
                            "$ref": "#/definitions/speaker"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/speakers/{id}/recordings": {
            "get": {
                "tags": [
                    "recordings"
                ],
                "description": "Fetch recordings by speaker id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Speaker id",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found recordings",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/recording"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "recording": {
            "type": "object",
            "required": [
                "recordingId",
                "recordingTime"
            ],
            "properties": {
                "recordingId": {
                    "type": "string",
                    "format": "uuid"
                },
                "deviceId": {
                    "type": "string",
                    "format": "uuid"
                },
                "speakerId": {
                    "type": "string",
                    "format": "uuid"
                },
                "settingsId": {
                    "type": "string",
                    "format": "uuid"
                },
                "recordingTime": {
                    "type": "string",
                    "format": "date-time"
                },
                "data": {
                    "type": "object"
                }
            }
        },
        "speaker": {
            "type": "object",
            "required": [
                "speakerId",
                "createdDate"
            ],
            "properties": {
                "speakerId": {
                    "type": "string",
                    "format": "uuid"
                },
                "deviceId": {
                    "type": "string",
                    "format": "uuid"
                },
                "createdDate": {
                    "type": "string",
                    "format": "date-time"
                },
                "data": {
                    "type": "object"
                }
            }
        }
    }
}